pyproject.toml
[project]
name = "multiagent-azure"
version = "0.1.0"
description = "Azure-first multi-agent orchestrator scaffold"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.2"
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0"
]
